requirements.txt
pyautogui==0.9.54
pillow==10.4.0
tkcap==0.0.3
pydicom==2.4.4
img2pdf==0.5.1
opencv_python==4.10.0.84
pandas==2.2.2
tensorflow==2.15.0
python-xlib==0.33
flask
grpcio
grpcio-tools
fastapi==0.112.0
pydantic==2.8
uvicorn==0.30.6